Review 1 fix error codes

+++ controllers/cardControllers.js
@@ -1,6 +1,6 @@
 const Card = require("../models/cardModel");
 
-const { ErrorNotFound } = require("../errors/ErrorNotFound");
+const {ErrorNotFound} = require("../errors/ErrorNotFound");
 
 const DefaultErrorCode = 500;
 const ErrorNotFoundCode = 404;
@@ -8,15 +8,15 @@ const ValidationErrorCode = 400;
 
 module.exports.getCards = (req, res) => {
   Card.find({})
-    .then((cards) => res.send({ data: cards }))
+    .then((cards) => res.send({data: cards}))
     .catch(() => {
-      res.status(DefaultErrorCode).send({ message: "Ошибка по умолчанию" });
+      res.status(DefaultErrorCode).send({message: "Ошибка по умолчанию"});
     });
 };
 
 module.exports.createCard = (req, res) => {
-  const { name, link } = req.body;
-  Card.create({ name, link, owner: req.user._id }).then((card) => res.send({
+  const {name, link} = req.body;
+  Card.create({name, link, owner: req.user._id}).then((card) => res.send({
     name: card.name,
     link: card.link,
     owner: card.owner,
@@ -25,74 +25,72 @@ module.exports.createCard = (req, res) => {
   }))
     .catch((err) => {
       if (err.name === "ValidationError") {
-        res.status(ValidationErrorCode).send({ message: "Переданы некорректные данные при создании карточки" });
+        res.status(ValidationErrorCode).send({message: "Переданы некорректные данные при создании карточки"});
       } else {
-        res.status(DefaultErrorCode).send({ message: "Ошибка по умолчанию" });
+        res.status(DefaultErrorCode).send({message: "Ошибка по умолчанию"});
       }
     });
 };
 
 module.exports.deleteCard = (req, res) => {
   Card.findById(req.params.cardId)
-    .orFail(() => {
-      ErrorNotFound("NotFoundError");
-    })
     .then((card) => {
-      if (card.owner.toString() === req.user._id) {
-        Card.findByIdAndDelete(req.params.cardId).then(() => res.send({
-          name: card.name,
-          link: card.link,
-          owner: card.owner,
-          likes: card.likes,
-          _id: card._id,
-        }));
+      if (card) {
+        if (card.owner.toString() === req.user._id) {
+          Card.findByIdAndDelete(req.params.cardId).then(() => res.send({
+            name: card.name,
+            link: card.link,
+            owner: card.owner,
+            likes: card.likes,
+            _id: card._id,
+          }));
+        }
+      } else {
+        res.status(ErrorNotFoundCode).send({message: "Карточка с указанным _id не найдена"});
       }
     })
     .catch((err) => {
       if (err.name === "CastError") {
-        res.status(ValidationErrorCode).send({ message: "Переданы некорректные данные удаления карточки" });
-      } else if (err.name === "DocumentNotFoundError") {
-        res.status(ErrorNotFoundCode).send({ message: "Карточка с указанным _id не найдена" });
+        res.status(ValidationErrorCode).send({message: "Переданы некорректные данные удаления карточки"});
       } else {
-        res.status(DefaultErrorCode).send({ message: "Ошибка по умолчанию" });
+        res.status(DefaultErrorCode).send({message: "Ошибка по умолчанию"});
       }
     });
 };
 
 module.exports.likeCard = (req, res) => {
-  Card.findByIdAndUpdate(req.params.cardId, { $addToSet: { likes: req.user._id } })
-    .orFail(() => {
-      ErrorNotFound("NotFoundError");
-    })
+  Card.findByIdAndUpdate(req.params.cardId, {$addToSet: {likes: req.user._id}})
     .then((card) => {
-      res.send({ data: card });
+      if (card) {
+        res.send({data: card});
+      } else {
+        res.status(ErrorNotFoundCode).send({message: "Передан несуществующий _id карточки"});
+      }
     })
     .catch((err) => {
       if (err.name === "CastError") {
-        res.status(ValidationErrorCode).send({ message: "Переданы некорректные данные для постановки лайка" });
-      } else if (err.name === "DocumentNotFoundError") {
-        res.status(ErrorNotFoundCode).send({ message: "Передан несуществующий _id карточки" });
+        res.status(ValidationErrorCode).send({message: "Переданы некорректные данные для постановки лайка"});
       } else {
-        res.status(DefaultErrorCode).send({ message: "Ошибка по умолчанию" });
+        res.status(DefaultErrorCode).send({message: "Ошибка по умолчанию"});
       }
     });
 };
 
 module.exports.unlikeCard = (req, res) => {
-  Card.findByIdAndUpdate(req.params.cardId, { $pull: { likes: req.user._id } })
-    .orFail(() => {
-      ErrorNotFound("NotFoundError");
-    })
+  Card.findByIdAndUpdate(req.params.cardId, {$pull: {likes: req.user._id}})
     .then((card) => {
-      res.send({ data: card });
+      if (card) {
+        res.send({data: card});
+      }
+      else{
+        res.status(ErrorNotFoundCode).send({message: "Переданы некорректные данные для удалении лайка"});
+      }
     })
     .catch((err) => {
       if (err.name === "CastError") {
-        res.status(ValidationErrorCode).send({ message: "Передан несуществующий _id карточки" });
-      } else if (err.name === "DocumentNotFoundError") {
-        res.status(ErrorNotFoundCode).send({ message: "Переданы некорректные данные для удалении лайка" });
+        res.status(ValidationErrorCode).send({message: "Передан несуществующий _id карточки"});
       } else {
-        res.status(DefaultErrorCode).send({ message: "Ошибка по умолчанию" });
+        res.status(DefaultErrorCode).send({message: "Ошибка по умолчанию"});
       }
     });
 };
